+++ targets/TARGET_STM/TARGET_STM32F0/common_objects.h
@@ -113,6 +113,15 @@ struct i2c_s {
 
 #include "gpio_object.h"
 
+#if DEVICE_ANALOGOUT
+struct dac_s {
+    DACName dac;
+    PinName pin;
+    uint32_t channel;
+    DAC_HandleTypeDef handle;
+};
+#endif
+
 #ifdef __cplusplus
 }
 #endif
