+++ targets/TARGET_STM/TARGET_STM32F3/TARGET_STM32F303x8/objects.h
@@ -60,13 +60,6 @@ struct analogin_s {
     uint32_t channel;
 };
 
-struct dac_s {
-    DACName dac;
-    PinName pin;
-    uint32_t channel;
-    DAC_HandleTypeDef handle;
-};
-
 struct can_s {
     CANName can;
     int index;
