+++ targets/TARGET_STM/TARGET_STM32F3/common_objects.h
@@ -111,6 +111,13 @@ struct i2c_s {
 #endif
 };
 
+struct dac_s {
+    DACName dac;
+    PinName pin;
+    uint32_t channel;
+    DAC_HandleTypeDef handle;
+};
+
 #include "gpio_object.h"
 
 #ifdef __cplusplus
