+++ targets/TARGET_STM/TARGET_STM32F4/TARGET_STM32F429xI/objects.h
@@ -60,11 +60,6 @@ struct analogin_s {
     uint8_t channel;
 };
 
-struct dac_s {
-    DACName dac;
-    uint8_t channel;
-};
-
 struct can_s {
     CANName can;
     int index;
