+++ targets/TARGET_STM/TARGET_STM32F7/common_objects.h
@@ -114,6 +114,13 @@ struct i2c_s {
 #define GPIO_IP_WITHOUT_BRR
 #include "gpio_object.h"
 
+struct dac_s {
+    DACName dac;
+    PinName pin;
+    uint32_t channel;
+    DAC_HandleTypeDef handle;
+};
+
 #ifdef __cplusplus
 }
 #endif
