+++ targets/TARGET_STM/TARGET_STM32L0/TARGET_DISCO_L053C8/objects.h
@@ -60,12 +60,6 @@ struct analogin_s {
     uint32_t channel;
 };
 
-struct dac_s {
-    DACName dac;
-    PinName pin;
-    uint32_t channel;
-};
-
 struct trng_s {
     RNG_HandleTypeDef handle;
 };
