+++ targets/TARGET_STM/TARGET_STM32L0/TARGET_NUCLEO_L053R8/objects.h
@@ -60,12 +60,6 @@ struct analogin_s {
     uint32_t channel;
 };
 
-struct dac_s {
-    DACName dac;
-    PinName pin;
-    uint32_t channel;
-};
-
 #include "common_objects.h"
 
 #ifdef __cplusplus
