+++ targets/TARGET_STM/TARGET_STM32L0/common_objects.h
@@ -118,6 +118,15 @@ struct flash_s {
 
 #include "gpio_object.h"
 
+#if DEVICE_ANALOGOUT
+struct dac_s {
+    DACName dac;
+    PinName pin;
+    uint32_t channel;
+    DAC_HandleTypeDef handle;
+};
+#endif
+
 #ifdef __cplusplus
 }
 #endif
