+++ targets/TARGET_STM/TARGET_STM32L1/TARGET_MOTE_L152RC/objects.h
@@ -60,11 +60,6 @@ struct analogin_s {
     uint32_t channel;
 };
 
-struct dac_s {
-    DACName dac;
-    PinName pin;
-};
-
 #define GPIO_IP_WITHOUT_BRR
 #include "common_objects.h"
