+++ targets/TARGET_STM/TARGET_STM32L1/TARGET_NZ32_SC151/objects.h
@@ -60,11 +60,6 @@ struct analogin_s {
     uint32_t channel;
 };
 
-struct dac_s {
-    DACName dac;
-    PinName pin;
-};
-
 #include "common_objects.h"
 
 #ifdef __cplusplus
