+++ targets/TARGET_STM/TARGET_STM32L4/TARGET_STM32L486xG/objects.h
@@ -60,12 +60,6 @@ struct analogin_s {
     uint32_t channel;
 };
 
-struct dac_s {
-    DACName dac;
-    PinName pin;
-    uint32_t channel;
-};
-
 struct can_s {
     CANName can;
     int index;
