+++ targets/TARGET_STM/TARGET_STM32L4/common_objects.h
@@ -118,6 +118,13 @@ struct flash_s {
 
 #include "gpio_object.h"
 
+struct dac_s {
+    DACName dac;
+    PinName pin;
+    uint32_t channel;
+    DAC_HandleTypeDef handle;
+};
+
 #ifdef __cplusplus
 }
 #endif
